// src/pages/create-agent/types.ts
export interface Agent {
  _id: string;
  name: string;
  description: string;
  agent_role: string;
  agent_instructions: string;
  features: Array<{
    type: string;
    config: Record<string, any>;
    priority: number;
  }>;
  tools?: string[];
  provider_id: string;
  model: string;
  temperature: string | number;
  top_p: string | number;
  tool?: string;
  tool_usage_description?: string;
  created_at?: string;
  updated_at?: string;
  llm_credential_id?: string;
  examples?: string;
  response_format:
    | { type: "json_schema"; json_schema: { strict: boolean; schema: any } }
    | { type: "text" };
}